requirements.txt
Flask
flask-cors
scikit-learn
requests
pandas
psycopg2-binary
python-dotenv